+++ src/main/java/miyucomics/hexical/casting/patterns/lesser_sentinel/OpLesserSentinelSet.kt
@@ -1,0 +1,36 @@
+package miyucomics.hexical.casting.patterns.lesser_sentinel
+
+import at.petrak.hexcasting.api.casting.RenderedSpell
+import at.petrak.hexcasting.api.casting.castables.SpellAction
+import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
+import at.petrak.hexcasting.api.casting.getList
+import at.petrak.hexcasting.api.casting.iota.Iota
+import at.petrak.hexcasting.api.casting.iota.Vec3Iota
+import at.petrak.hexcasting.api.casting.mishaps.MishapBadCaster
+import at.petrak.hexcasting.api.casting.mishaps.MishapInvalidIota
+import at.petrak.hexcasting.api.misc.MediaConstants
+import miyucomics.hexical.interfaces.PlayerEntityMinterface
+import net.minecraft.server.network.ServerPlayerEntity
+import net.minecraft.util.math.Vec3d
+
+class OpLesserSentinelSet : SpellAction {
+	override val argc = 1
+	override fun execute(args: List<Iota>, env: CastingEnvironment): SpellAction.Result {
+		if (env.castingEntity !is ServerPlayerEntity)
+			throw MishapBadCaster()
+		val positions = args.getList(0, argc).map {
+			if (it !is Vec3Iota)
+				throw MishapInvalidIota.of(args[0], 0, "lesser_sentinel_list")
+			it.vec3
+		}
+		return SpellAction.Result(Spell(positions), 0, listOf())
+	}
+
+	private data class Spell(val pos: List<Vec3d>) : RenderedSpell {
+		override fun cast(env: CastingEnvironment) {
+			val lesserSentinels = (env.castingEntity as PlayerEntityMinterface).getLesserSentinels()
+			lesserSentinels.lesserSentinels = pos.toMutableList()
+			lesserSentinels.syncToClient(env.castingEntity as ServerPlayerEntity)
+		}
+	}
+}