+++ src/main/java/miyucomics/hexical/client/ClientStorage.kt
@@ -1,9 +1,11 @@
 package miyucomics.hexical.client
 
 import miyucomics.hexical.data.LedgerInstance
+import net.minecraft.util.math.Vec3d
 
 object ClientStorage {
 	@JvmField
-	var time: Int = 0
+	var ticks: Int = 0
 	var ledger: LedgerInstance = LedgerInstance()
+	var lesserSentinels: MutableList<Vec3d> = mutableListOf()
 }